优化

+++ app/src/main/java/com/moe/video/framework/util/FileUtil.java
@@ -1,15 +1,14 @@
 package com.moe.video.framework.util;
 import java.io.File;
 
-public class FileUtil
-{
-	public static void delete(File f){
-		if(f.isFile())
+public class FileUtil {
+	public static void delete(File f) {
+		if (f.isFile())
 			f.delete();
-			else{
-		for(File ff:f.listFiles())
-			delete(ff);
+        else if (f.isDirectory()) {
+            for (File ff:f.listFiles())
+                delete(ff);
 			f.delete();
-			}
+        }
 	}
 }
